requirements.txt
torch
transformers
pillow
pytesseract
pdf2image
opencv-python
torchvision
docling_core